+++ src/pages/LetterWritePage/components/LetterReceiverBottomSheet.tsx
@@ -1,9 +1,10 @@
 import { useState } from 'react';
 import { useFormContext } from 'react-hook-form';
+import { css } from '@emotion/react';
+import COLORS from '@/constants/colors';
+import textStyles from '@/styles/textStyles';
 import BottomSheet from '@/components/BottomSheet';
 import { ArrowClockWise } from '@/assets/icons';
-import COLORS from '@/constants/colors';
-import style from '../styles';
 import { BottomSheetProps } from './LetterWriteContent';
 import { AgeSlider, GenderSelect, ConcernSelect } from '.';
 
@@ -60,3 +61,46 @@ const LetterReceiverSelect = ({
 };
 
 export default LetterReceiverSelect;
+
+const style = {
+  bottomSheetContainer: css`
+    padding-inline: 1rem;
+  `,
+  bottomSheetTitle: (rotaion: number) => css`
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    height: 2.75rem;
+
+    h2 {
+      ${textStyles.t2};
+    }
+
+    svg {
+      cursor: pointer;
+      transition: transform 0.5s;
+      transform: rotate(${rotaion}deg);
+    }
+  `,
+  buttonContainer: css`
+    display: flex;
+
+    button {
+      cursor: pointer;
+      flex: 1;
+      padding: 0.75rem 1.25rem;
+      border: 0;
+      ${textStyles.t3}
+    }
+
+    & > button:nth-of-type(1) {
+      background-color: ${COLORS.gray5};
+      color: ${COLORS.gray1};
+    }
+
+    & > button:nth-of-type(2) {
+      background-color: ${COLORS.primary};
+      color: white;
+    }
+  `,
+};
