+++ src/pages/LetterWritePage/components/index.tsx
@@ -1,8 +1,11 @@
-import AgeSlider from './AgeSlider';
-import ConcernSelect from './ConcernSelect';
-import GenderSelect from './GenderSelect';
+import AgeSlider from './inputs/AgeSlider';
+import ConcernSelect from './inputs/ConcernSelect';
+import GenderSelect from './inputs/GenderSelect';
+import LetterTextarea from './inputs/LetterTextarea';
+import ImageSelect from './inputs/ImageSelect';
 import LetterPaper from './LetterPaper';
-import LetterReceiverSelect from './LetterReceiverSelect';
+import PolaroidImage from './PolaroidImage';
+import LetterReceiverBottomSheet from './LetterReceiverBottomSheet';
 import LetterReceiverContainer from './LetterReceiverContainer';
 import LetterWriteBottom from './LetterWriteBottom';
 import LetterWriteContent from './LetterWriteContent';
@@ -11,8 +14,11 @@ export {
   AgeSlider,
   ConcernSelect,
   GenderSelect,
+  LetterTextarea,
+  ImageSelect,
   LetterPaper,
-  LetterReceiverSelect,
+  PolaroidImage,
+  LetterReceiverBottomSheet,
   LetterReceiverContainer,
   LetterWriteBottom,
   LetterWriteContent,
